Added an area of effect object
Added area of effect as well as damage over time to that area

+++ Projects/2DSmallPrototype/Assets/Scripts/Enemy.cs
@@ -2,19 +2,19 @@ using UnityEngine;
 
 public class Enemy : MonoBehaviour
 {
-    public int maxHealth = 100;
-    private int currentHealth;
+    public float maxHealth = 100f;
+    private float currentHealth;
 
     void Start()
     {
         currentHealth = maxHealth;
     }
 
-    public void TakeDamage(int damage)
+    public void TakeDamage(float damage)
     {
         currentHealth -= damage;
 
-        if (currentHealth <= 0)
+        if (currentHealth <= 0f)
         {
             Die();
         }
@@ -33,7 +33,7 @@ public class Enemy : MonoBehaviour
 
             if (bullet != null)
             {
-                TakeDamage(bullet.damage);
+                TakeDamage(bullet.damage); // Bullet.damage can stay int, will convert to float
             }
 
             Destroy(other.gameObject);
